+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Each change should be on one line, as GitHub markdown in the release section can
 - `metalarchives` plugin: fetches lyrics from metal-archives.com (Encyclopaedia Metallum).
 - `songlyrics` plugin: fetches lyrics from songlyrics.com.
 - `utatime` plugin: fetches lyrics from utatime.com (formerly lyrical-nonsense.com).
+- `utanet` plugin: fetches lyrics from uta-net.com (Uta-Net).
 
 ## [1.0.0] - 2026-08-04